+++ examples/tools/tool-package-quickstart/my_tool_package/tools/tool_with_file_path_input.py
@@ -4,10 +4,7 @@ from promptflow import tool
 from promptflow.contracts.types import FilePath
 
 
-@tool(
-    name="Tool with FilePath Input",
-    description="This is a tool to demonstrate the usage of FilePath input",
-)
+@tool
 def my_tool(input_file: FilePath, input_text: str) -> str:
     # customise your own code to handle and use the input_file here
     new_module = importlib.import_module(Path(input_file).stem)