+++ src/promptflow/promptflow/_sdk/entities/_run.py
@@ -5,7 +5,6 @@
 import datetime
 import functools
 import json
-import logging
 import uuid
 from os import PathLike
 from pathlib import Path
@@ -20,7 +19,6 @@ from promptflow._sdk._constants import (
     DEFAULT_VARIANT,
     FLOW_DIRECTORY_MACRO_IN_CONFIG,
     FLOW_RESOURCE_ID_PREFIX,
-    LOGGER_NAME,
     PARAMS_OVERRIDE_KEY,
     PROMPT_FLOW_DIR_NAME,
     REGISTRY_URI_PREFIX,
@@ -48,6 +46,7 @@ from promptflow._sdk._utils import (
 from promptflow._sdk.entities._yaml_translatable import YAMLTranslatableMixin
 from promptflow._sdk.schemas._run import RunSchema
 from promptflow._utils.flow_utils import get_flow_lineage_id
+from promptflow._utils.logger_utils import get_cli_sdk_logger
 from promptflow.exceptions import UserErrorException
 
 AZURE_RUN_TYPE_2_RUN_TYPE = {
@@ -61,6 +60,9 @@ REST_RUN_TYPE_2_RUN_TYPE = {
     RestRunTypes.EVALUATION: RunTypes.EVALUATION,
     RestRunTypes.PAIRWISE_EVALUATE: RunTypes.PAIRWISE_EVALUATE,
 }
+
+
+logger = get_cli_sdk_logger()
 
 
 class Run(YAMLTranslatableMixin):
@@ -150,6 +152,7 @@ class Run(YAMLTranslatableMixin):
         # init here to make sure those fields initialized in all branches.
         self.flow = flow
         self._use_remote_flow = is_remote_uri(flow)
+        # What's this?
         self._experiment_name = None
         self._lineage_id = None
         if self._use_remote_flow:
@@ -414,6 +417,8 @@ class Run(YAMLTranslatableMixin):
         params_override: Optional[list] = None,
         **kwargs,
     ):
+        from marshmallow import INCLUDE
+
         data = data or {}
         params_override = params_override or []
         context = {
@@ -424,6 +429,7 @@ class Run(YAMLTranslatableMixin):
             data=data,
             context=context,
             additional_message="Failed to load flow run",
+            unknown=INCLUDE,
             **kwargs,
         )
         if yaml_path:
@@ -487,6 +493,7 @@ class Run(YAMLTranslatableMixin):
         from promptflow.azure._restclient.flow.models import (
             BatchDataInput,
             RunDisplayNameGenerationType,
+            SessionSetupModeEnum,
             SubmitBulkRunRequest,
         )
 
@@ -522,6 +529,18 @@ class Run(YAMLTranslatableMixin):
                         )
                     inputs_mapping[k] = val
 
+        # parse resources
+        if self._resources is not None:
+            if not isinstance(self._resources, dict):
+                raise TypeError(f"resources should be a dict, got {type(self._resources)} for {self._resources}")
+            vm_size = self._resources.get("instance_type", None)
+            max_idle_time_minutes = self._resources.get("idle_time_before_shutdown_minutes", None)
+            # change to seconds
+            max_idle_time_seconds = max_idle_time_minutes * 60 if max_idle_time_minutes else None
+        else:
+            vm_size = None
+            max_idle_time_seconds = None
+
         # use functools.partial to avoid too many arguments that have the same values
         common_submit_bulk_run_request = functools.partial(
             SubmitBulkRunRequest,
@@ -541,6 +560,9 @@ class Run(YAMLTranslatableMixin):
             connections=self.connections,
             flow_lineage_id=self._lineage_id,
             run_display_name_generation_type=RunDisplayNameGenerationType.USER_PROVIDED_MACRO,
+            vm_size=vm_size,
+            max_idle_time_seconds=max_idle_time_seconds,
+            session_setup_mode=SessionSetupModeEnum.SYSTEM_WAIT,
         )
 
         if str(self.flow).startswith(REMOTE_URI_PREFIX):
@@ -636,7 +658,7 @@ class Run(YAMLTranslatableMixin):
                     f"{config.get_run_output_path()!r}; "
                     f"will use default output path: {path!r} instead."
                 )
-                logging.getLogger(LOGGER_NAME).warning(warning_message)
+                logger.warning(warning_message)
         return (path / str(self.name)).resolve()
 
     @classmethod