+++ src/promptflow/promptflow/_sdk/operations/_local_storage_operations.py
@@ -12,7 +12,6 @@ from functools import partial
 from pathlib import Path
 from typing import Any, Dict, List, NewType, Optional, Tuple, Union
 
-import yaml
 from filelock import FileLock
 
 from promptflow._sdk._constants import (
@@ -23,7 +22,7 @@ from promptflow._sdk._constants import (
     PROMPT_FLOW_DIR_NAME,
     LocalStorageFilenames,
 )
-from promptflow._sdk._errors import BulkRunException
+from promptflow._sdk._errors import BulkRunException, InvalidRunError
 from promptflow._sdk._utils import PromptflowIgnoreFile, generate_flow_tools_json
 from promptflow._sdk.entities import Run
 from promptflow._sdk.entities._flow import Flow
@@ -31,6 +30,7 @@ from promptflow._utils.dataclass_serializer import serialize
 from promptflow._utils.exception_utils import PromptflowExceptionPresenter
 from promptflow._utils.logger_utils import LogContext, get_cli_sdk_logger
 from promptflow._utils.multimedia_utils import get_file_reference_encoder
+from promptflow._utils.yaml_utils import load_yaml
 from promptflow.batch._result import BatchResult
 from promptflow.contracts.multimedia import Image
 from promptflow.contracts.run_info import FlowRunInfo
@@ -213,6 +213,12 @@ class LocalStorageOperations(AbstractRunStorage):
 
         self._dump_meta_file()
 
+    def delete(self) -> None:
+        def on_rmtree_error(func, path, exc_info):
+            raise InvalidRunError(f"Failed to delete run {self.path} due to {exc_info[1]}.")
+
+        shutil.rmtree(path=self.path, onerror=on_rmtree_error)
+
     def _dump_meta_file(self) -> None:
         with open(self._meta_path, mode="w", encoding=DEFAULT_ENCODING) as f:
             json.dump({"batch_size": LOCAL_STORAGE_BATCH_SIZE}, f, ensure_ascii=False)
@@ -246,7 +252,7 @@ class LocalStorageOperations(AbstractRunStorage):
     def load_io_spec(self) -> Tuple[Dict[str, Dict[str, str]], Dict[str, Dict[str, str]]]:
         """Load input/output spec from DAG."""
         with open(self._dag_path, mode="r", encoding=DEFAULT_ENCODING) as f:
-            flow_dag = yaml.safe_load(f)
+            flow_dag = load_yaml(f)
         return flow_dag["inputs"], flow_dag["outputs"]
 
     def load_inputs(self) -> RunInputs: