+++ src/promptflow/promptflow/executor/_async_nodes_scheduler.py
@@ -9,6 +9,7 @@ import os
 import signal
 import threading
 import time
+import traceback
 from asyncio import Task
 from concurrent.futures import ThreadPoolExecutor
 from typing import Any, Dict, List, Tuple
@@ -16,10 +17,14 @@ from typing import Any, Dict, List, Tuple
 from promptflow._core.flow_execution_context import FlowExecutionContext
 from promptflow._core.tools_manager import ToolsManager
 from promptflow._utils.logger_utils import flow_logger
-from promptflow._utils.utils import set_context
+from promptflow._utils.utils import extract_user_frame_summaries, set_context
 from promptflow.contracts.flow import Node
 from promptflow.executor._dag_manager import DAGManager
 from promptflow.executor._errors import NoNodeExecutedError
+
+PF_ASYNC_NODE_SCHEDULER_EXECUTE_TASK_NAME = "_pf_async_nodes_scheduler.execute"
+DEFAULT_TASK_LOGGING_INTERVAL = 60
+ASYNC_DAG_MANAGER_COMPLETED = False
 
 
 class AsyncNodesScheduler:
@@ -31,6 +36,9 @@ class AsyncNodesScheduler:
         self._tools_manager = tools_manager
         # TODO: Add concurrency control in execution
         self._node_concurrency = node_concurrency
+        self._task_start_time = {}
+        self._task_last_log_time = {}
+        self._dag_manager_completed_event = threading.Event()
 
     async def execute(
         self,
@@ -38,8 +46,26 @@ class AsyncNodesScheduler:
         inputs: Dict[str, Any],
         context: FlowExecutionContext,
     ) -> Tuple[dict, dict]:
-        signal.signal(signal.SIGINT, signal_handler)
-        signal.signal(signal.SIGTERM, signal_handler)
+        # TODO: Provide cancel API
+        if threading.current_thread() is threading.main_thread():
+            signal.signal(signal.SIGINT, signal_handler)
+            signal.signal(signal.SIGTERM, signal_handler)
+        else:
+            flow_logger.info(
+                "Current thread is not main thread, skip signal handler registration in AsyncNodesScheduler."
+            )
+
+        loop = asyncio.get_running_loop()
+        monitor = threading.Thread(
+            target=monitor_long_running_coroutine,
+            args=(loop, self._task_start_time, self._task_last_log_time, self._dag_manager_completed_event),
+            daemon=True,
+        )
+        monitor.start()
+
+        # Set the name of scheduler tasks to avoid monitoring its duration
+        task = asyncio.current_task()
+        task.set_name(PF_ASYNC_NODE_SCHEDULER_EXECUTE_TASK_NAME)
 
         parent_context = contextvars.copy_context()
         executor = ThreadPoolExecutor(
@@ -67,6 +93,9 @@ class AsyncNodesScheduler:
             task2nodes = await self._wait_and_complete_nodes(task2nodes, dag_manager)
             submitted_tasks2nodes = self._execute_nodes(dag_manager, context, executor)
             task2nodes.update(submitted_tasks2nodes)
+        # Set the event to notify the monitor thread to exit
+        # Ref: https://docs.python.org/3/library/threading.html#event-objects
+        self._dag_manager_completed_event.set()
         for node in dag_manager.bypassed_nodes:
             dag_manager.completed_nodes_outputs[node] = None
         return dag_manager.completed_nodes_outputs, dag_manager.bypassed_nodes
@@ -75,6 +104,8 @@ class AsyncNodesScheduler:
         if not task2nodes:
             raise NoNodeExecutedError("No nodes are ready for execution, but the flow is not completed.")
         tasks = [task for task in task2nodes]
+        for task in tasks:
+            self._task_start_time[task] = time.time()
         done, _ = await asyncio.wait(tasks, return_when=asyncio.FIRST_COMPLETED)
         dag_manager.complete_nodes({task2nodes[task].name: task.result() for task in done})
         for task in done:
@@ -135,6 +166,104 @@ def signal_handler(sig, frame):
     monitor = threading.Thread(target=monitor_coroutine_after_cancellation, args=(loop,))
     monitor.start()
     raise KeyboardInterrupt
+
+
+def log_stack_recursively(task: asyncio.Task, elapse_time: float):
+    """Recursively log the frame of a task or coroutine.
+    Traditional stacktrace would stop at the first awaited nested inside the coroutine.
+
+    :param task: Task to log
+    :type task_or_coroutine: asyncio.Task
+    :param elapse_time: Seconds elapsed since the task started
+    :type elapse_time: float
+    """
+    # We cannot use task.get_stack() to get the stack, because only one stack frame is
+    # returned for a suspended coroutine because of the implementation of CPython
+    # Ref: https://github.com/python/cpython/blob/main/Lib/asyncio/tasks.py
+    # "only one stack frame is returned for a suspended coroutine."
+    task_or_coroutine = task
+    frame_summaries = []
+    # Collect frame_summaries along async call chain
+    while True:
+        if isinstance(task_or_coroutine, asyncio.Task):
+            # For a task, get the coroutine it's running
+            coroutine: asyncio.coroutine = task_or_coroutine.get_coro()
+        elif asyncio.iscoroutine(task_or_coroutine):
+            coroutine = task_or_coroutine
+        else:
+            break
+
+        frame = coroutine.cr_frame
+        stack_summary: traceback.StackSummary = traceback.extract_stack(frame)
+        frame_summaries.extend(stack_summary)
+        task_or_coroutine = coroutine.cr_await
+
+    # Format the frame summaries to warning message
+    if frame_summaries:
+        user_frame_summaries = extract_user_frame_summaries(frame_summaries)
+        stack_messages = traceback.format_list(user_frame_summaries)
+        all_stack_message = "".join(stack_messages)
+        task_msg = (
+            f"Task {task.get_name()} has been running for {elapse_time:.0f} seconds,"
+            f" stacktrace:\n{all_stack_message}"
+        )
+        flow_logger.warning(task_msg)
+
+
+def monitor_long_running_coroutine(
+    loop: asyncio.AbstractEventLoop,
+    task_start_time: dict,
+    task_last_log_time: dict,
+    dag_manager_completed_event: threading.Event,
+):
+    flow_logger.info("monitor_long_running_coroutine started")
+
+    logging_interval = DEFAULT_TASK_LOGGING_INTERVAL
+    logging_interval_in_env = os.environ.get("PF_TASK_PEEKING_INTERVAL")
+    if logging_interval_in_env:
+        try:
+            value = int(logging_interval_in_env)
+            if value <= 0:
+                raise ValueError
+            logging_interval = value
+            flow_logger.info(
+                f"Using value of PF_TASK_PEEKING_INTERVAL in environment variable as "
+                f"logging interval: {logging_interval_in_env}"
+            )
+        except ValueError:
+            flow_logger.warning(
+                f"Value of PF_TASK_PEEKING_INTERVAL in environment variable ('{logging_interval_in_env}') "
+                f"is invalid, use default value {DEFAULT_TASK_LOGGING_INTERVAL}"
+            )
+
+    while not dag_manager_completed_event.is_set():
+        running_tasks = [task for task in asyncio.all_tasks(loop) if not task.done()]
+        # get duration of running tasks
+        for task in running_tasks:
+            # Do not monitor the scheduler task
+            if task.get_name() == PF_ASYNC_NODE_SCHEDULER_EXECUTE_TASK_NAME:
+                continue
+            # Do not monitor sync tools, since they will run in executor thread and will
+            # be monitored by RepeatLogTimer.
+            task_stacks = task.get_stack()
+            if (
+                task_stacks
+                and task_stacks[-1].f_code
+                and task_stacks[-1].f_code.co_name == AsyncNodesScheduler._sync_function_to_async_task.__name__
+            ):
+                continue
+            if task_start_time.get(task) is None:
+                flow_logger.warning(f"task {task.get_name()} has no start time, which should not happen")
+            else:
+                duration = time.time() - task_start_time[task]
+                if duration > logging_interval:
+                    if (
+                        task_last_log_time.get(task) is None
+                        or time.time() - task_last_log_time[task] > logging_interval
+                    ):
+                        log_stack_recursively(task, duration)
+                        task_last_log_time[task] = time.time()
+        time.sleep(1)
 
 
 def monitor_coroutine_after_cancellation(loop: asyncio.AbstractEventLoop):