+++ src/promptflow/tests/sdk_cli_azure_test/conftest.py
@@ -26,8 +26,10 @@ from .recording_utilities import (
     PFAzureIntegrationTestRecording,
     SanitizedValues,
     VariableRecorder,
+    get_created_flow_name_from_flow_path,
     get_pf_client_for_replay,
     is_live,
+    is_record,
     is_replay,
 )
 
@@ -164,12 +166,14 @@ def flow_serving_client_remote_connection(mocker: MockerFixture, remote_workspac
 
 
 @pytest.fixture
-def flow_serving_client_with_prt_config_env(mocker: MockerFixture, subscription_id, resource_group_name, workspace_name):  # noqa: E501
+def flow_serving_client_with_prt_config_env(
+    mocker: MockerFixture, subscription_id, resource_group_name, workspace_name
+):  # noqa: E501
     connections = {
-            "PRT_CONFIG_OVERRIDE": f"deployment.subscription_id={subscription_id},"
-            f"deployment.resource_group={resource_group_name},"
-            f"deployment.workspace_name={workspace_name},"
-            "app.port=8088",
+        "PRT_CONFIG_OVERRIDE": f"deployment.subscription_id={subscription_id},"
+        f"deployment.resource_group={resource_group_name},"
+        f"deployment.workspace_name={workspace_name},"
+        "app.port=8088",
     }
     return create_serving_client_with_connections("basic-with-connection", mocker, connections)
 
@@ -191,7 +195,7 @@ def flow_serving_client_with_aml_resource_id_env(mocker: MockerFixture, remote_w
 def serving_client_with_connection_name_override(mocker: MockerFixture, remote_workspace_resource_id):
     connections = {
         "aoai_connection": "azure_open_ai_connection",
-        "PROMPTFLOW_CONNECTION_PROVIDER": remote_workspace_resource_id
+        "PROMPTFLOW_CONNECTION_PROVIDER": remote_workspace_resource_id,
     }
     return create_serving_client_with_connections("llm_connection_override", mocker, connections)
 
@@ -209,9 +213,7 @@ def serving_client_with_connection_data_override(mocker: MockerFixture, remote_w
     return create_serving_client_with_connections(model_name, mocker, connections)
 
 
-def create_serving_client_with_connections(
-    model_name, mocker: MockerFixture, connections: dict = {}
-):
+def create_serving_client_with_connections(model_name, mocker: MockerFixture, connections: dict = {}):
     from promptflow._sdk._serving.app import create_app as create_serving_app
 
     model_path = (Path(MODEL_ROOT) / model_name).resolve().absolute().as_posix()
@@ -352,7 +354,7 @@ def mock_get_user_identity_info(mocker: MockerFixture) -> None:
 
 
 @pytest.fixture(scope=package_scope_in_live_mode())
-def created_flow(pf: PFClient, randstr: Callable[[str], str]) -> Flow:
+def created_flow(pf: PFClient, randstr: Callable[[str], str], variable_recorder: VariableRecorder) -> Flow:
     """Create a flow for test."""
     flow_display_name = randstr("flow_display_name")
     flow_source = FLOWS_DIR + "/simple_hello_world/"
@@ -368,7 +370,15 @@ def created_flow(pf: PFClient, randstr: Callable[[str], str]) -> Flow:
     assert result.display_name == flow_display_name
     assert result.type == FlowType.STANDARD
     assert result.tags == tags
-    assert result.path.endswith(f"/promptflow/{flow_display_name}/flow.dag.yaml")
+    assert result.path.endswith("flow.dag.yaml")
+
+    # flow in Azure will have different file share name with timestamp
+    # and this is a client-side behavior, so we need to sanitize this in recording
+    # so extract this during record test
+    if is_record():
+        flow_name_const = "flow_name"
+        flow_name = get_created_flow_name_from_flow_path(result.path)
+        variable_recorder.get_or_record_variable(flow_name_const, flow_name)
 
     yield result
 
@@ -488,3 +498,14 @@ def mock_isinstance_for_mock_datastore() -> None:
 
         with patch("builtins.isinstance", new=mock_isinstance):
             yield
+
+
+@pytest.fixture(autouse=True)
+def mock_check_latest_version() -> None:
+    """Mock check latest version.
+
+    As CI uses docker, it will always trigger this check behavior, and we don't have recording for this;
+    and this will hit many unknown issue with vcrpy.
+    """
+    with patch("promptflow._utils.version_hint_utils.check_latest_version", new=lambda: None):
+        yield