+++ src/promptflow/CHANGELOG.md
@@ -7,16 +7,27 @@
 - [Executor] Calculate system_metrics recursively in api_calls.
 - [Executor] Add flow root level api_calls, so that user can overview the aggregated metrics of a flow.
 - [Executor] Add @trace decorator to make it possible to log traces for functions that are called by tools.
+- [SDK/CLI][azure] Switch automatic runtime's session provision to system wait.
+- [SDK/CLI] Add `--skip-open-browser` option to `pf flow serve` to skip opening browser.
+- [SDK/CLI][azure] Support submit flow to sovereign cloud.
+- [SDK/CLI] Support `pf run delete` to delete a run irreversibly.
 
 ### Bugs Fixed
 
 - Fix unaligned inputs & outputs or pandas exception during get details against run in Azure.
 - Fix loose flow path validation for run schema.
 - Fix "Without Import Data" in run visualize page results from invalid JSON value (`-Infinity`, `Infinity` and `NaN`).
+- Fix "ValueError: invalid width -1" when show-details against long column(s) in narrow terminal window.
+- Fix invalid tool code generated when initializing the script tool with icon.
 
 ### Improvements
 
-- [SDK/CLI] For `pfazure flow create` used by non-msft tenant user, use user name instead of user object id in the remote flow folder path. (e.g. `Users/<user-name>/promptflow`).
+- [SDK/CLI] For `pfazure flow create`:
+  - If used by non-msft tenant user, use user name instead of user object id in the remote flow folder path. (e.g. `Users/<user-name>/promptflow`).
+  - When flow has unknown attributes, log warning instead of raising error.
+  - Use local flow folder name and timestamp as the azure flow file share folder name.
+- [SDK/CLI] For `pf/pfazure run create`, when run has unknown attribute, log warning instead of raising error.
+- Replace `pyyaml` with `ruamel.yaml` to adopt YAML 1.2 specification.
 
 ## 1.3.0 (2023.12.27)